refactor: Add DatabaseService - Updated BreedsRepository to delegate all local data operations to DatabaseService
