+++ CatsApp/Data/Repository/BreedsRepository.swift
@@ -22,9 +22,11 @@ struct PageResult {
 
 struct BreedsRepository: BreedsRepositoryProtocol {
     private let service: BreedsFetching
+    private let database: DatabaseServiceProtocol
 
-    init(service: BreedsFetching = BreedsDataService()) {
+    init(service: BreedsFetching = BreedsDataService(), database: DatabaseServiceProtocol = DatabaseService()) {
         self.service = service
+        self.database = database
     }
 
     @MainActor
@@ -35,7 +37,7 @@ struct BreedsRepository: BreedsRepositoryProtocol {
                 limit: limit
             )
             let count = dtos.count
-            let saved = try await savePage(dtos, context: context)
+            let saved = try database.saveBreeds(dtos, context: context)
             return PageResult(breeds: saved, fetchedCount: count)
         } catch let error as DomainError {
             throw error
@@ -45,87 +47,26 @@ struct BreedsRepository: BreedsRepositoryProtocol {
     }
 
     @MainActor
-    func savePage(_ dtos: [BreedsDataService.CatBreed], context: ModelContext)
-        async throws -> [CatBreed]
-    {
-        guard !dtos.isEmpty else { return try fetchAll(context: context) }
-
-        let existingDescriptor = FetchDescriptor<CatBreed>()
-        
-        let existing: [CatBreed]
+    func savePage(_ dtos: [BreedsDataService.CatBreed], context: ModelContext) async throws -> [CatBreed] {
         do {
-            existing = try context.fetch(existingDescriptor)
+            return try database.saveBreeds(dtos, context: context)
         } catch {
-            throw DomainError.persistenceError
-        }
-        
-        var existingIds = Set(existing.map { $0.id })
-
-        var inserted = false
-        for dto in dtos {
-            if existingIds.contains(dto.id) { continue }
-            let model = CatBreed(
-                id: dto.id,
-                name: dto.name,
-                origin: dto.origin,
-                temperament: dto.temperament,
-                breedDescription: dto.description,
-                lifeSpan: dto.life_span,
-                referenceImageId: dto.reference_image_id,
-                isFavorite: false
-            )
-            context.insert(model)
-            existingIds.insert(dto.id)
-            inserted = true
-        }
-        if inserted {
-            do {
-                try context.save()
-            } catch {
-                throw DomainError.persistenceError
-            }
-        }
-
-        do{
-            let all = try context.fetch(existingDescriptor)
-            return CatBreed.sortedByName(all)
-        } catch {
-            throw DomainError.persistenceError
+            throw error
         }
     }
 
     @MainActor
     func toggleFavorite(_ breed: CatBreed, context: ModelContext) throws {
-        breed.isFavorite.toggle()
-        do {
-            try context.save()
-        } catch {
-            throw DomainError.persistenceError
-        }
-         
+        try database.toggleFavorite(breed, context: context)
     }
 
     @MainActor
     func fetchFavorites(context: ModelContext) throws -> [CatBreed] {
-        let predicate = #Predicate<CatBreed> { $0.isFavorite == true }
-        let descriptor = FetchDescriptor<CatBreed>(predicate: predicate)
-        
-        do {
-            let favorites = try context.fetch(descriptor)
-            return CatBreed.sortedByName(favorites)
-        } catch {
-            throw DomainError.persistenceError
-        }
+        try database.fetchFavoriteBreeds(context: context)
     }
 
     @MainActor
     func fetchAll(context: ModelContext) throws -> [CatBreed] {
-        let descriptor = FetchDescriptor<CatBreed>()
-        do{
-            let all = try context.fetch(descriptor)
-            return CatBreed.sortedByName(all)
-        } catch{
-            throw DomainError.persistenceError
-        }
+        try database.fetchAllBreeds(context: context)
     }
 }
